New small manifests!
Thanks to Ulf accepting and implementing my feature request in
9a4ebdd, we now have small `A`-indexed singletons as data definitions.

One reason why I wanted to allow a large `A` is that, by using a
pointed type, we can already have an extensible type of labels!

Still, my attempts at using ".." for making the `A` or labels
irrelevant in the correct way are not working properly.